Update Flights-Within-X-Miles.py
This script retrieves information about aircraft from a local flight tracker
(PiAware with FlightAware Pro Stick)
and calculates the distance between the plane's coordinates and a set of home
coordinates. If the plane is within 3 miles of the home coordinates and at an
altitude of 3000 feet or lower, the flight information is printed.
The script continuously runs in an infinite loop, with a 2-second pause between
each iteration. 
The script makes use of the requests library to retrieve data from a URL and the
geopy library to calculate the geodesic distance between two sets of coordinates.